Abstract storyBookWrapper for reuse
