+++ src/utils/storyBookWrapper.js
@@ -1,0 +1,11 @@
+import styled from 'react-emotion';
+
+const Wrapper = styled('div')`
+  height: 100vh;
+  width: 100vw;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+`;
+
+export default Wrapper;
